[Puja|Fadli] Modify migration script to check existing row before change the column

+++ db/migrate/20190627134000_remove_image_from_infrastructure_components.rb
@@ -1,5 +1,13 @@
 class RemoveImageFromInfrastructureComponents < ActiveRecord::Migration[5.2]
-  def change
-    remove_column :infrastructure_components, :image, :string
+  def up
+    if InfrastructureComponent.where.not(image: nil).empty?
+      remove_column :infrastructure_components, :image, :string
+    else
+      raise $!, "There is infrastructure component that has image value"
+    end
+  end
+
+  def down
+    add_column :infrastructure_components, :image, :string
   end
 end
